+++ packages/bigcommerce-payments-integration/src/bigcommerce-payments-fastlane/bigcommerce-payments-fastlane-payment-strategy.ts
@@ -456,6 +456,7 @@ export default class BigCommercePaymentsFastlanePaymentStrategy implements Payme
 
             if (authenticationState === TDSecureAuthenticationState.Cancelled) {
                 console.error('3DS check was canceled');
+                throw new PaymentMethodInvalidError();
             }
         }