+++ packages/braintree-utils/src/index.ts
@@ -7,6 +7,7 @@ export { default as BraintreeMessages } from './braintree-messages';
 export { default as BraintreeScriptLoader } from './braintree-script-loader';
 export { default as BraintreeSdk } from './braintree-sdk';
 export { default as createBraintreeSdk } from './create-braintree-sdk';
+export { default as BraintreeSDKVersionManager } from './braintree-sdk-version-manager';
 export { BRAINTREE_SDK_SCRIPTS_INTEGRITY } from './braintree-sdk-scripts-integrity';
 export { BRAINTREE_SDK_STABLE_VERSION } from './braintree-sdk-verison';