+++ packages/core/src/payment/strategies/braintree/create-braintree-payment-processor.ts
@@ -1,13 +1,24 @@
 import { ScriptLoader } from '@bigcommerce/script-loader';
 
-import { BraintreeScriptLoader } from '@bigcommerce/checkout-sdk/braintree-utils';
+import {
+    BraintreeScriptLoader,
+    BraintreeSDKVersionManager,
+} from '@bigcommerce/checkout-sdk/braintree-utils';
+import { PaymentIntegrationService } from '@bigcommerce/checkout-sdk/payment-integration-api';
 
 import BraintreeHostedForm from './braintree-hosted-form';
 import BraintreePaymentProcessor from './braintree-payment-processor';
 import BraintreeSDKCreator from './braintree-sdk-creator';
 
-export default function createBraintreePaymentProcessor(scriptLoader: ScriptLoader) {
-    const braintreeScriptLoader = new BraintreeScriptLoader(scriptLoader, window);
+export default function createBraintreePaymentProcessor(
+    scriptLoader: ScriptLoader,
+    paymentIntegrationService: PaymentIntegrationService,
+) {
+    const braintreeScriptLoader = new BraintreeScriptLoader(
+        scriptLoader,
+        window,
+        new BraintreeSDKVersionManager(paymentIntegrationService),
+    );
     const braintreeSDKCreator = new BraintreeSDKCreator(braintreeScriptLoader);
     const braintreeHostedForm = new BraintreeHostedForm(braintreeSDKCreator);