+++ packages/paypal-commerce-utils/src/utils/index.ts
@@ -3,3 +3,4 @@ export { default as isPayPalFastlaneCustomer } from './is-paypal-fastlane-custom
 export { default as getFastlaneStyles } from './get-fastlane-styles';
 export { default as getPaypalMessagesStylesFromBNPLConfig } from './get-paypal-messages-styles-from-bnpl-config';
 export { default as isRedirectActionError } from './is-redirect-action-error';
+export { default as isPaypalFastlaneRequestError } from './is-paypal-fastlane-request-error';