+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.774.0](https://github.com/bigcommerce/checkout-sdk-js/compare/v1.773.0...v1.774.0) (2025-07-23)
+
+
+### Features
+
+* **payment:** added BraintreeSDKVersionManager ([#2932](https://github.com/bigcommerce/checkout-sdk-js/issues/2932)) ([afa2f19](https://github.com/bigcommerce/checkout-sdk-js/commit/afa2f195a74cfe266d67183a8e47060c1e033b9f))
+
+## [1.773.0](https://github.com/bigcommerce/checkout-sdk-js/compare/v1.772.0...v1.773.0) (2025-07-23)
+
+
+### Features
+
+* **payment:** Added corresponding error for 422 payment status ([#2938](https://github.com/bigcommerce/checkout-sdk-js/issues/2938)) ([31c3813](https://github.com/bigcommerce/checkout-sdk-js/commit/31c3813dcca0861a73ffb9e6e9a4559dd15427aa))
+
+## [1.772.0](https://github.com/bigcommerce/checkout-sdk-js/compare/v1.771.0...v1.772.0) (2025-07-23)
+
+
+### Features
+
+* **payment:** Create order after liability shift passed when 3ds is on ([#2936](https://github.com/bigcommerce/checkout-sdk-js/issues/2936)) ([51ddb2e](https://github.com/bigcommerce/checkout-sdk-js/commit/51ddb2ee24c603fb9cf99170a132bd6c62e14205))
+
+
+### Bug Fixes
+
+* **customer:** Fetch updated checkout after carts are merged ([#2923](https://github.com/bigcommerce/checkout-sdk-js/issues/2923)) ([b5c94c4](https://github.com/bigcommerce/checkout-sdk-js/commit/b5c94c424a52d1df62e77730aa724cc31a3c4622))
+
 ## [1.771.0](https://github.com/bigcommerce/checkout-sdk-js/compare/v1.770.0...v1.771.0) (2025-07-22)